consolidate run_helpers.py & helpers.py bugfix for get_model_path
